% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz-aliases.R
\name{estimate_rician_sigma}
\alias{estimate_rician_sigma}
\title{Preferred alias for estimate_sigma_rician}
\usage{
estimate_rician_sigma(vec, mask = NULL)
}
\arguments{
\item{vec}{4D fMRI data as a numeric array or `neuroim2::NeuroVec`.}

\item{mask}{Optional 3D logical/0-1 mask (array or `NeuroVol`). If `NULL`,
all voxels are used.}
}
\description{
Preferred alias for estimate_sigma_rician
}
\seealso{
estimate_sigma_rician
}
