% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz-aliases.R
\name{guided_filter3d}
\alias{guided_filter3d}
\title{Preferred alias for st_guided_filter3d}
\usage{
guided_filter3d(vol, radius = 4L, eps = 0.7^2, guide = NULL, mask = NULL)
}
\arguments{
\item{vol}{3D array or neuroim2::NeuroVol}

\item{radius}{integer window radius (voxels)}

\item{eps}{regularization (variance floor) in signal units^2}

\item{guide}{3D array or NeuroVol; if NULL uses \code{vol}}

\item{mask}{3D logical/0-1 mask; NULL => all voxels}
}
\description{
Preferred alias for st_guided_filter3d
}
\seealso{
st_guided_filter3d
}
