% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_guided_filter.R
\name{st_guided_filter3d}
\alias{st_guided_filter3d}
\title{3D edge-preserving guided filter}
\usage{
st_guided_filter3d(vol, radius = 4L, eps = 0.7^2, guide = NULL, mask = NULL)
}
\arguments{
\item{vol}{3D array or neuroim2::NeuroVol}

\item{radius}{integer window radius (voxels)}

\item{eps}{regularization (variance floor) in signal units^2}

\item{guide}{3D array or NeuroVol; if NULL uses \code{vol}}

\item{mask}{3D logical/0-1 mask; NULL => all voxels}
}
\value{
same type as input (NeuroVol if given), otherwise 3D array
}
\description{
3D edge-preserving guided filter
}
