% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz-aliases.R
\name{vst_forward}
\alias{vst_forward}
\title{Preferred alias for vst_forward_rician}
\usage{
vst_forward(x, sigma)
}
\arguments{
\item{x}{Numeric array (3D/4D) of magnitudes.}

\item{sigma}{Noise standard deviation.}
}
\description{
Preferred alias for vst_forward_rician
}
\seealso{
vst_forward_rician
}
