% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz-aliases.R
\name{vst_inverse}
\alias{vst_inverse}
\title{Preferred alias for vst_inverse_rician}
\usage{
vst_inverse(z, sigma)
}
\arguments{
\item{z}{Transformed array (output of `vst_forward_rician`).}

\item{sigma}{Noise standard deviation.}
}
\description{
Preferred alias for vst_inverse_rician
}
\seealso{
vst_inverse_rician
}
